+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2023-10-03
+- Add IsMod and IsSub to Chatter class
+
 ## [1.2.1] - 2023-09-01
 
 ### Added
 - Leave Channel method
-- IsBroadcastter added to chatter class
+- IsBroadcaster added to chatter class
 
 ## [1.2.0] - 2023-08-15